removed Tx msgs that are not Lexus LS specific
Was getting tmux a error "opendbc_repo/opendbc/can/packer.cc: undefined address 0"

Maybe the CAN message(s) that are created in toyotacan.py are not defined in the Lexus LS dbc file. So, that's why keep getting this error.

Doesn't hurt to remove since only Lexus LS specific message are needed to be sent.